// apps/desktop/src/types/generated/SubmissionRef.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SubmissionProvider } from "./SubmissionProvider";
import type { SubmissionType } from "./SubmissionType";

export type SubmissionRef = {
  provider: SubmissionProvider;
  submissionType: SubmissionType;
  submissionId: string;
};